test: move from-import test to test_unsupported.py
Moved test_from_import_function_calls_unresolvable from integration
tests to unit tests in test_unsupported.py since it documents a
temporary limitation rather than testing core functionality.

The test now uses analyze_source() directly to avoid I/O operations
and maintain proper separation between unit and integration tests.
